#3 První vylepšení správce médií Odstranění dialogu pro import média. Nyní se otevře galerie rovnou
