+++ app/src/main/java/cz/zcu/fav/remotestimulatorcontrol/ui/configurations/detail/ConfigurationDetailActivity.java
@@ -1,11 +1,19 @@
 package cz.zcu.fav.remotestimulatorcontrol.ui.configurations.detail;
 
+import android.Manifest;
+import android.annotation.TargetApi;
 import android.content.Intent;
+import android.content.pm.PackageManager;
 import android.databinding.DataBindingUtil;
+import android.net.Uri;
+import android.os.Build;
 import android.os.Bundle;
 import android.os.Handler;
 import android.os.Message;
+import android.support.annotation.NonNull;
 import android.support.design.widget.Snackbar;
+import android.support.v4.app.ActivityCompat;
+import android.support.v4.content.ContextCompat;
 import android.support.v7.app.ActionBar;
 import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.LinearLayoutManager;
@@ -38,6 +46,7 @@ import cz.zcu.fav.remotestimulatorcontrol.ui.configurations.detail.fvep.Configur
 import cz.zcu.fav.remotestimulatorcontrol.ui.configurations.detail.mediaimport.MediaImportActivity;
 import cz.zcu.fav.remotestimulatorcontrol.ui.configurations.detail.rea.ConfigurationFragmentREA;
 import cz.zcu.fav.remotestimulatorcontrol.ui.configurations.detail.tvep.ConfigurationFragmentTVEP;
+import cz.zcu.fav.remotestimulatorcontrol.util.FileUtils;
 import cz.zcu.fav.remotestimulatorcontrol.widget.editableseekbar.EditableSeekBar;
 
 public class ConfigurationDetailActivity extends AppCompatActivity
@@ -55,7 +64,8 @@ public class ConfigurationDetailActivity extends AppCompatActivity
     @SuppressWarnings("unused")
     private static final String TAG = "ConfigDetailActivity";
     private static final String FRAGMENT = "fragment";
-    private static final int REQUEST_PATH = 1;
+    private static final int REQUEST_FILE_PATH = 1;
+    private static final int REQUEST_PERMISSION = 2;
     // endregion
 
     // region Variables
@@ -79,6 +89,8 @@ public class ConfigurationDetailActivity extends AppCompatActivity
     private int id;
     // Použitý typ souboru
     private ExtensionType extensionType;
+    // Příznak určující, zda-li mi uživatel povolil přístup k souborovému systému
+    private boolean permissionGranted = false;
     // endregion
 
     // region Private methods
@@ -148,6 +160,15 @@ public class ConfigurationDetailActivity extends AppCompatActivity
         setResult(RESULT_OK, intent);
         finish();
     }
+
+    @TargetApi(Build.VERSION_CODES.JELLY_BEAN)
+    private void verifyPermission() {
+        if (ContextCompat.checkSelfPermission(this, Manifest.permission.READ_EXTERNAL_STORAGE) != PackageManager.PERMISSION_GRANTED) {
+            ActivityCompat.requestPermissions(this, new String[]{Manifest.permission.READ_EXTERNAL_STORAGE}, REQUEST_PERMISSION);
+        } else {
+            permissionGranted = true;
+        }
+    }
     // endregion
 
     @Override
@@ -185,6 +206,8 @@ public class ConfigurationDetailActivity extends AppCompatActivity
         if (actionBar != null) {
             actionBar.setDisplayHomeAsUpEnabled(true);
         }
+
+        verifyPermission();
     }
 
     @Override
@@ -210,9 +233,18 @@ public class ConfigurationDetailActivity extends AppCompatActivity
     @Override
     protected void onActivityResult(int requestCode, int resultCode, Intent data) {
         switch (requestCode) {
-            case REQUEST_PATH:
+            case REQUEST_FILE_PATH:
                 if (resultCode == RESULT_OK) {
-                    String path = data.getStringExtra(MediaImportActivity.MEDIA_PATH);
+                    Uri uri = data.getData();
+                    String path = FileUtils.getPath(this, uri);
+
+                    if (path == null) {
+                        Snackbar.make(recyclerView, "Fail", Snackbar.LENGTH_SHORT).show();
+                        Log.e(TAG, "Nepodařilo se naimportovat externí médium.");
+                        Log.e(TAG, uri.toString());
+                        return;
+                    }
+
                     mediaManager.importt(new File(path));
                 }
                 break;
@@ -249,6 +281,18 @@ public class ConfigurationDetailActivity extends AppCompatActivity
         initRecyclerView();
     }
 
+    @Override
+    public void onRequestPermissionsResult(int requestCode, @NonNull String[] permissions, @NonNull int[] grantResults) {
+        switch (requestCode) {
+            case REQUEST_PERMISSION:
+                Log.d(TAG, "Byl vynesen request na práva");
+                if (grantResults.length > 0 && grantResults[0] == PackageManager.PERMISSION_GRANTED) {
+                    permissionGranted = true;
+                }
+                break;
+        }
+    }
+
     @SuppressWarnings("unused")
     public void onMediaCheckBoxClick(View view) {
         boolean checked = ((CheckBox) view).isChecked();
@@ -270,7 +314,17 @@ public class ConfigurationDetailActivity extends AppCompatActivity
 
     @Override
     public void onAddMediaClick() {
-        startActivityForResult(new Intent(this, MediaImportActivity.class), REQUEST_PATH);
+//        startActivityForResult(new Intent(this, MediaImportActivity.class), REQUEST_PATH);
+        if (!permissionGranted) {
+            Log.i(TAG, "Nebyla udělena práva pro přístup k úložišti");
+            verifyPermission();
+            return;
+        }
+
+        Intent intent = new Intent();
+        intent.setAction(Intent.ACTION_GET_CONTENT);
+        intent.setType("image/*, audio/*");
+        startActivityForResult(Intent.createChooser(intent, "Vyberte externí médium"), REQUEST_FILE_PATH);
     }
 
     // region MediaManager handler
