--- app/src/main/res/layout/activity_media_import.xml
@@ -1,69 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<layout>
-
-    <data>
-
-        <import type="android.view.View" />
-
-        <import type="cz.zcu.fav.remotestimulatorcontrol.ui.configurations.detail.mediaimport.MediaImportActivity" />
-
-        <import type="cz.zcu.fav.remotestimulatorcontrol.ui.configurations.detail.mediaimport.ObservableMedia" />
-
-        <variable
-            name="controler"
-            type="MediaImportActivity" />
-
-        <variable
-            name="media"
-            type="ObservableMedia" />
-
-    </data>
-
-    <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-        xmlns:tools="http://schemas.android.com/tools"
-        android:id="@+id/activity_media_import"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:orientation="vertical"
-        tools:context="cz.zcu.fav.remotestimulatorcontrol.ui.configurations.detail.mediaimport.MediaImportActivity">
-
-        <TextView
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:paddingLeft="8dp"
-            android:paddingRight="8dp"
-            android:text="@string/file_path" />
-
-        <Button
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:text="@{media.path}"
-            android:onClick="onFilePathRequest"/>
-
-        <LinearLayout
-            style="?android:actionBarStyle"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:orientation="horizontal">
-
-            <Button
-                android:id="@+id/buttonCancel"
-                style="?android:actionButtonStyle"
-                android:layout_width="0dp"
-                android:layout_height="wrap_content"
-                android:layout_weight="1"
-                android:onClick="onCancel"
-                android:text="@android:string/cancel" />
-
-            <Button
-                android:id="@+id/buttonImport"
-                style="?android:actionButtonStyle"
-                android:layout_width="0dp"
-                android:layout_height="wrap_content"
-                android:layout_weight="1"
-                android:enabled="@{media.changed &amp;&amp; media.validityFlag == 0}"
-                android:onClick="onImport"
-                android:text="@android:string/ok" />
-        </LinearLayout>
-    </LinearLayout>
-</layout>
